refactor(xdc): remove RewardsStore, compute rewards on demand
Eliminates the RewardsStore persistence layer entirely. Epoch rewards
are now calculated on the fly via XdcEpochRewardCalculator (new singleton,
no ITransactionProcessor dependency) and returned directly through the
IXdcRewardCalculator interface used by XdcRpcModule.GetRewardByAccount.

XdcRewardCalculator becomes a thin wrapper that delegates computation to
XdcEpochRewardCalculator and calls IMintedRecordContract.UpdateAccounting
on the TIP-upgrade path. The XdcRewardsDb column family and its
IRewardsStore abstraction are removed. Tests are updated accordingly.
