+++ src/Nethermind/Nethermind.Xdc/XdcEpochRewardCalculator.cs
@@ -1,0 +1,347 @@
+// SPDX-FileCopyrightText: 2025 Demerzel Solutions Limited
+// SPDX-License-Identifier: LGPL-3.0-only
+
+using Nethermind.Blockchain;
+using Nethermind.Consensus.Rewards;
+using Nethermind.Core;
+using Nethermind.Core.Crypto;
+using Nethermind.Core.Specs;
+using Nethermind.Crypto;
+using Nethermind.Int256;
+using Nethermind.Xdc.Contracts;
+using Nethermind.Xdc.Spec;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Nethermind.Xdc;
+
+/// <summary>
+/// Calculates XDC epoch rewards on demand without persisting results or mutating state.
+/// </summary>
+/// <remarks>
+/// Rewards are paid only at epoch checkpoints. Pre-upgrade: proportional split of
+/// <see cref="IXdcReleaseSpec.Reward"/> across masternode signatures. TIP-upgrade: fixed
+/// per-signer rewards for masternode/protector/observer. Holder split is always 90% owner,
+/// 10% foundation.
+/// </remarks>
+public class XdcEpochRewardCalculator(
+    IEpochSwitchManager epochSwitchManager,
+    ISpecProvider specProvider,
+    IBlockTree blockTree,
+    IMasternodeVotingContract masternodeVotingContract,
+    ISigningTxCache signingTxCache) : IXdcRewardCalculator
+{
+    private readonly EthereumEcdsa _ethereumEcdsa = new(specProvider.ChainId);
+
+    public BlockReward[] CalculateEpochRewards(XdcBlockHeader epochHeader)
+        => CalculateEpochRewardsCore(epochHeader).Rewards;
+
+    /// <summary>
+    /// Calculates epoch rewards and returns accounting values for the minted-record contract update.
+    /// </summary>
+    /// <remarks>
+    /// <c>IsTipUpgrade</c> indicates whether the TIP-upgrade code path was taken; only when
+    /// <c>true</c> should the caller invoke <c>IMintedRecordContract.UpdateAccounting</c>.
+    /// </remarks>
+    internal (BlockReward[] Rewards, UInt256 BurnedInEpoch, bool IsTipUpgrade) CalculateEpochRewardsCore(XdcBlockHeader epochHeader)
+    {
+        ArgumentNullException.ThrowIfNull(epochHeader);
+        if (epochHeader.Number == 0)
+            return ([], UInt256.Zero, false);
+
+        if (!epochSwitchManager.IsEpochSwitchAtBlock(epochHeader))
+            return ([], UInt256.Zero, false);
+
+        ulong number = epochHeader.Number;
+        IXdcReleaseSpec spec = specProvider.GetXdcSpec(epochHeader, epochHeader.ExtraConsensusData.BlockRound);
+        if (number == spec.SwitchBlock + 1)
+            return ([], UInt256.Zero, false);
+
+        Address foundationWalletAddr = spec.FoundationWallet;
+        if (foundationWalletAddr == default || foundationWalletAddr == Address.Zero)
+            throw new InvalidOperationException("Foundation wallet address cannot be empty");
+
+        UInt256 totalFoundationWalletReward = UInt256.Zero;
+        List<BlockReward> rewards = [];
+        (Dictionary<Address, ulong> masternodeSigners, Dictionary<Address, ulong> protectorSigners, Dictionary<Address, ulong> observerSigners, UInt256 burnedInOneEpoch) = GetSigningTxCount(epochHeader, spec);
+
+        if (!spec.IsTipUpgradeRewardEnabled)
+        {
+            UInt256 chainReward = (UInt256)spec.Reward * Unit.Ether;
+            Dictionary<Address, UInt256> rewardSigners = CalculateRewardForSigners(chainReward, masternodeSigners);
+            AddDistributedRewards(epochHeader, foundationWalletAddr, rewardSigners, rewards, ref totalFoundationWalletReward);
+        }
+        else
+        {
+            Dictionary<Address, UInt256> masternodeRewards = CalculateFixedRewardForSigners(spec.MasternodeReward, masternodeSigners);
+            Dictionary<Address, UInt256> protectorRewards = CalculateFixedRewardForSigners(spec.ProtectorReward, protectorSigners);
+            Dictionary<Address, UInt256> observerRewards = CalculateFixedRewardForSigners(spec.ObserverReward, observerSigners);
+
+            AddDistributedRewards(epochHeader, foundationWalletAddr, masternodeRewards, rewards, ref totalFoundationWalletReward);
+            AddDistributedRewards(epochHeader, foundationWalletAddr, protectorRewards, rewards, ref totalFoundationWalletReward);
+            AddDistributedRewards(epochHeader, foundationWalletAddr, observerRewards, rewards, ref totalFoundationWalletReward);
+        }
+
+        if (totalFoundationWalletReward > UInt256.Zero)
+            rewards.Add(new BlockReward(foundationWalletAddr, totalFoundationWalletReward));
+
+        return ([.. rewards], burnedInOneEpoch, spec.IsTipUpgradeRewardEnabled);
+    }
+
+    private (
+        Dictionary<Address, ulong> MasternodeSigners,
+        Dictionary<Address, ulong> ProtectorSigners,
+        Dictionary<Address, ulong> ObserverSigners,
+        UInt256 BurnedInOneEpoch) GetSigningTxCount(XdcBlockHeader epochHeader, IXdcReleaseSpec spec)
+    {
+        Dictionary<Address, ulong> masternodeSigners = [];
+        Dictionary<Address, ulong> protectorSigners = [];
+        Dictionary<Address, ulong> observerSigners = [];
+        UInt256 burnedInOneEpoch = UInt256.Zero;
+        ulong number = epochHeader.Number;
+        if (number == 0) return (masternodeSigners, protectorSigners, observerSigners, burnedInOneEpoch);
+
+        ulong signEpochCount = 1, rewardEpochCount = 2, epochCount = 0, endBlockNumber = 0, startBlockNumber = 0;
+
+        Dictionary<ulong, Hash256> blockNumberToHash = [];
+        Dictionary<Hash256, HashSet<Address>> hashToSigningAddress = [];
+        HashSet<Address> masternodes = [];
+        HashSet<Address> protectors = [];
+        HashSet<Address> observers = [];
+        ulong mergeSignRange = spec.MergeSignRange;
+
+        XdcBlockHeader h = epochHeader;
+        ulong blockIdx = number - 1;
+        while (true)
+        {
+            Hash256 parentHash = h.ParentHash;
+            h = (XdcBlockHeader)blockTree.FindHeader(parentHash!, blockIdx) ?? throw new InvalidOperationException($"Header with hash {parentHash} not found");
+            if (epochCount == 0 && !h.BaseFeePerGas.IsZero)
+            {
+                UInt256 burnedInBlock = h.BaseFeePerGas * (UInt256)h.GasUsed;
+                burnedInOneEpoch += burnedInBlock;
+            }
+            if (epochSwitchManager.IsEpochSwitchAtBlock(h) && h.Number != spec.SwitchBlock + 1)
+            {
+                epochCount++;
+                if (epochCount == signEpochCount) endBlockNumber = blockIdx;
+                if (epochCount == rewardEpochCount)
+                {
+                    startBlockNumber = blockIdx + 1;
+                    masternodes = GetRewardMasternodes(h, spec);
+
+                    if (spec.IsTipUpgradeRewardEnabled)
+                    {
+                        // TIPUpgradeReward path: select protector and observer sets from stake-sorted candidates.
+                        // Exclude current masternodes and penalized nodes from the checkpoint header.
+                        Address[] candidatesByStake = GetCandidatesByStakeForReward(epochHeader);
+                        int penaltiesCount = h.PenaltiesAddress?.Length ?? 0;
+                        HashSet<Address> excludedCandidates = new(masternodes.Count + penaltiesCount);
+                        excludedCandidates.UnionWith(masternodes);
+                        if (h.PenaltiesAddress is not null)
+                            excludedCandidates.UnionWith(h.PenaltiesAddress);
+
+                        foreach (Address candidate in candidatesByStake)
+                        {
+                            if (candidate == Address.Zero || excludedCandidates.Contains(candidate))
+                                continue;
+
+                            if (protectors.Count < spec.MaxProtectorNodes)
+                                protectors.Add(candidate);
+                            else if (observers.Count < spec.MaxObserverNodes)
+                                observers.Add(candidate);
+                        }
+                    }
+                    break;
+                }
+            }
+
+            blockNumberToHash[blockIdx] = h.Hash;
+            Transaction[] signingTxs = signingTxCache.GetSigningTransactions(h.Hash, blockIdx, spec);
+
+            foreach (Transaction tx in signingTxs)
+            {
+                Hash256 blockHash = ExtractBlockHashFromSigningTxData(tx.Data);
+                tx.SenderAddress ??= _ethereumEcdsa.RecoverAddress(tx);
+                if (!hashToSigningAddress.ContainsKey(blockHash))
+                    hashToSigningAddress[blockHash] = [];
+                hashToSigningAddress[blockHash].Add(tx.SenderAddress);
+            }
+
+            if (blockIdx == 0) break;
+            blockIdx--;
+        }
+
+        // Only blocks at heights that are multiples of MergeSignRange are considered.
+        // Calculate start >= startBlockNumber so that start % MergeSignRange == 0
+        ulong start = (startBlockNumber + mergeSignRange - 1) / mergeSignRange * mergeSignRange;
+        for (ulong i = start; i < endBlockNumber; i += mergeSignRange)
+        {
+            if (!blockNumberToHash.TryGetValue(i, out Hash256 blockHash)) continue;
+            if (!hashToSigningAddress.TryGetValue(blockHash, out HashSet<Address> addresses)) continue;
+            foreach (Address addr in addresses)
+            {
+                if (masternodes.Contains(addr))
+                    IncrementSignerCount(masternodeSigners, addr);
+                else if (protectors.Contains(addr))
+                    IncrementSignerCount(protectorSigners, addr);
+                else if (observers.Contains(addr))
+                    IncrementSignerCount(observerSigners, addr);
+            }
+        }
+        return (masternodeSigners, protectorSigners, observerSigners, burnedInOneEpoch);
+    }
+
+    protected internal virtual HashSet<Address> GetRewardMasternodes(XdcBlockHeader checkpointHeader, IXdcReleaseSpec spec)
+    {
+        if (checkpointHeader.Number <= spec.SwitchBlock)
+        {
+            return [.. checkpointHeader.ExtraData.ParseV1Masternodes()];
+        }
+
+        return [.. checkpointHeader.ValidatorsAddress!];
+    }
+
+    private Address[] GetCandidatesByStakeForReward(XdcBlockHeader checkpointHeader)
+    {
+        // We intentionally avoid GetCandidatesByStake here to preserve Go-equivalent ordering:
+        // fetch candidates at the checkpoint header and apply a stable stake-descending sort locally.
+        Address[] candidates = masternodeVotingContract.GetCandidates(checkpointHeader) ?? [];
+        if (candidates.Length == 0)
+            return [];
+
+        List<CandidateStake> candidatesAndStake = new(candidates.Length);
+        foreach (Address candidate in candidates)
+        {
+            if (candidate == Address.Zero)
+                continue;
+
+            candidatesAndStake.Add(new CandidateStake
+            {
+                Address = candidate,
+                Stake = masternodeVotingContract.GetCandidateStake(checkpointHeader, candidate),
+            });
+        }
+
+        return candidatesAndStake
+            .OrderByDescending(static x => x.Stake)
+            .Select(static x => x.Address)
+            .ToArray();
+    }
+
+    private static void IncrementSignerCount(Dictionary<Address, ulong> signers, Address addr)
+    {
+        if (!signers.TryAdd(addr, 1))
+            signers[addr] += 1;
+    }
+
+    private Hash256 ExtractBlockHashFromSigningTxData(ReadOnlyMemory<byte> data)
+    {
+        ReadOnlySpan<byte> span = data.Span;
+        if (span.Length != XdcConstants.SignTransactionDataLength)
+            throw new ArgumentException($"Signing tx calldata must be exactly {XdcConstants.SignTransactionDataLength} bytes.", nameof(data));
+
+        // 36..67: bytes32 blockHash
+        ReadOnlySpan<byte> hashBytes = span.Slice(36, 32);
+        return new Hash256(hashBytes);
+    }
+
+    private Dictionary<Address, UInt256> CalculateRewardForSigners(UInt256 totalReward,
+        Dictionary<Address, ulong> signers)
+    {
+        Dictionary<Address, UInt256> rewardSigners = [];
+        ulong totalSigningCount = 0;
+        foreach (ulong signerCount in signers.Values)
+            totalSigningCount += signerCount;
+
+        foreach ((Address signer, ulong count) in signers)
+        {
+            UInt256 reward = CalculateProportionalReward(count, totalSigningCount, totalReward);
+            rewardSigners.Add(signer, reward);
+        }
+        return rewardSigners;
+    }
+
+    private Dictionary<Address, UInt256> CalculateFixedRewardForSigners(
+        UInt256 rewardPerSigner,
+        Dictionary<Address, ulong> signers)
+    {
+        Dictionary<Address, UInt256> rewardSigners = [];
+        if (rewardPerSigner == UInt256.Zero)
+            return rewardSigners;
+
+        foreach ((Address signer, ulong _) in signers)
+            rewardSigners[signer] = rewardPerSigner;
+
+        return rewardSigners;
+    }
+
+    /// <summary>
+    /// Calculates a proportional reward based on the number of signatures.
+    /// Uses UInt256 arithmetic to maintain precision with large Wei values.
+    ///
+    /// Formula: (totalReward / totalSignatures) * signatureCount
+    /// </summary>
+    internal UInt256 CalculateProportionalReward(
+        ulong signatureCount,
+        ulong totalSignatures,
+        UInt256 totalReward)
+    {
+        if (signatureCount == 0 || totalSignatures == 0)
+        {
+            return UInt256.Zero;
+        }
+
+        UInt256 signatures = (UInt256)signatureCount;
+        UInt256 total = (UInt256)totalSignatures;
+
+        UInt256 portion = totalReward / total;
+        UInt256 reward = portion * signatures;
+
+        return reward;
+    }
+
+    /// <summary>
+    /// Distributes a signer's reward using the 90/10 owner/foundation split.
+    /// </summary>
+    /// <remarks>
+    /// If the masternode owner is the foundation wallet, the reference client stores only the
+    /// foundation entry (overwrite semantics), so the holder reward is the 10% foundation amount
+    /// and the separate foundation contribution is zero.
+    /// </remarks>
+    protected internal virtual (BlockReward HolderReward, UInt256 FoundationWalletReward) DistributeRewards(
+        XdcBlockHeader epochHeader, Address masternodeAddress, UInt256 reward, Address foundationWalletAddr)
+    {
+        Address owner = masternodeVotingContract.GetCandidateOwner(epochHeader, masternodeAddress);
+
+        // 90% of the reward goes to the masternode owner
+        UInt256 masterReward = reward * 90 / 100;
+
+        // 10% of the reward goes to the foundation wallet
+        UInt256 foundationReward = reward / 10;
+
+        // The reference client stores both entries in a map, so the foundation reward replaces the owner reward on collision.
+        if (owner == foundationWalletAddr)
+        {
+            return (new BlockReward(owner, foundationReward), UInt256.Zero);
+        }
+
+        return (new BlockReward(owner, masterReward), foundationReward);
+    }
+
+    private void AddDistributedRewards(
+        XdcBlockHeader epochHeader,
+        Address foundationWalletAddr,
+        Dictionary<Address, UInt256> rewardSigners,
+        List<BlockReward> rewards,
+        ref UInt256 totalFoundationWalletReward)
+    {
+        foreach ((Address signer, UInt256 reward) in rewardSigners)
+        {
+            (BlockReward holderReward, UInt256 foundationWalletReward) = DistributeRewards(epochHeader, signer, reward, foundationWalletAddr);
+            totalFoundationWalletReward += foundationWalletReward;
+            rewards.Add(holderReward);
+        }
+    }
+}
