+++ src/Nethermind/Nethermind.Xdc/XdcModule.cs
@@ -104,9 +104,9 @@ public class XdcModule : Module
             .AddSingleton<IEpochSwitchManager, EpochSwitchManager>()
             .AddSingleton<IXdcConsensusContext, XdcConsensusContext>()
             .AddDatabase(XdcRocksDbConfigFactory.XdcSnapshotDbName)
-            .AddDatabase(XdcRocksDbConfigFactory.XdcRewardsDbName)
             .AddSingleton<IPenaltyHandler, PenaltyHandler>()
-            .AddSingleton<IRewardsStore, RewardsStore>()
+            .AddSingleton<XdcEpochRewardCalculator>()
+            .Bind<IXdcRewardCalculator, XdcEpochRewardCalculator>()
             .AddSingleton<ITimeoutTimer, TimeoutTimer>()
             .AddSingleton<ISyncInfoManager, SyncInfoManager>()
