+++ src/Nethermind/Nethermind.Xdc/XdcRocksDbConfigFactory.cs
@@ -12,13 +12,11 @@ namespace Nethermind.Xdc;
 public class XdcRocksDbConfigFactory(IRocksDbConfigFactory baseFactory, IDbConfig dbConfig) : IRocksDbConfigFactory
 {
     public const string XdcSnapshotDbName = "XdcSnapshots";
-    public const string XdcRewardsDbName = "XdcRewards";
 
     public IRocksDbConfig GetForDatabase(string databaseName, string? columnName)
     {
-        // For XdcSnapshots, create PerTableDbConfig with validate=false since it doesn't have
-        // XdcSnapshotsDbRocksDbOptions configured in IDbConfig
-        if (databaseName is XdcSnapshotDbName or XdcRewardsDbName)
+        // XdcSnapshots doesn't have dedicated RocksDb options in IDbConfig, so bypass validation.
+        if (databaseName is XdcSnapshotDbName)
         {
             return new PerTableDbConfig(dbConfig, databaseName, columnName, validate: false);
         }
