+++ src/Nethermind/Nethermind.Xdc/XdcSubnetRewardCalculatorSource.cs
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: 2026 Demerzel Solutions Limited
 // SPDX-License-Identifier: LGPL-3.0-only
 
-using Nethermind.Blockchain;
 using Nethermind.Consensus.Rewards;
 using Nethermind.Core.Specs;
 using Nethermind.Evm.TransactionProcessing;
@@ -10,29 +9,11 @@ using Nethermind.Xdc.Contracts;
 namespace Nethermind.Xdc;
 
 internal sealed class XdcSubnetRewardCalculatorSource(
-    IEpochSwitchManager epochSwitchManager,
+    XdcSubnetRewardCalculator subnetEpochCalculator,
     ISpecProvider specProvider,
-    IBlockTree blockTree,
-    IMasternodeVotingContract masternodeVotingContract,
-    IMintedRecordContract mintedRecordContract,
-    ISigningTxCache signingTxCache,
-    IRewardsStore rewardsStore)
-    : XdcRewardCalculatorSource(
-        epochSwitchManager,
-        specProvider,
-        blockTree,
-        masternodeVotingContract,
-        mintedRecordContract,
-        signingTxCache,
-        rewardsStore)
+    IMintedRecordContract mintedRecordContract)
+    : XdcRewardCalculatorSource(subnetEpochCalculator, specProvider, mintedRecordContract)
 {
-    public override IRewardCalculator Get(ITransactionProcessor processor) => new XdcSubnetRewardCalculator(
-        EpochSwitchManager,
-        SpecProvider,
-        BlockTree,
-        MasternodeVotingContract,
-        MintedRecordContract,
-        SigningTxCache,
-        processor,
-        RewardsStore);
+    public override IRewardCalculator Get(ITransactionProcessor processor) =>
+        new XdcRewardCalculator(EpochRewardCalculator, SpecProvider, MintedRecordContract, processor);
 }
